Change model ID for hub models
The code was using a fixed modelID to ensure that jslm2 always used the
correct prompt regardless of path. That caused the wrong prompt to be
used for hub models though - this changes the ID to the hub model name,
so FastChat may use its normal method to determine the prompt to be
used. (Note that method may not work, but that's an existing limitation
of FastChat.)